test: Refactor group user reports tests
Remove the base test class UserReportEnvironmentTestCase since we only use
this in one place now.
This is a follow up to https://github.com/getsentry/sentry/pull/14293
